requirements.txt

streamlit
langchain
google-generativeai
localtunnel
